Add tslint warning for unused stuff. Remove unused stuff.
