+++ server/dfsType.ts
@@ -213,7 +213,7 @@ class DFSState implements beebfs.IFSState {
     public libDrive: string;
     public libDir: string;
 
-    private readonly log: utils.Log;
+    private readonly log: utils.Log;// tslint:disable-line no-unused-variable
 
     public constructor(volume: beebfs.Volume, settings: any | undefined, log: utils.Log) {
         this.volume = volume;
