+++ server/speedtest.ts
@@ -1,7 +1,6 @@
 import * as utils from './utils';
 import { BNL } from './utils';
 import * as crypto from 'crypto';
-import * as assert from 'assert';
 
 /////////////////////////////////////////////////////////////////////////
 /////////////////////////////////////////////////////////////////////////
